requirements.txt
requests>=2.31.0
click>=8.1.0
rich>=13.0.0
pyyaml>=6.0
tabulate>=0.9.0
pycryptodome>=3.19.0 